Create web3 address utils
Move the `isSameEthAddress` fn to the `web3/utils/address` and add
`isAddrress` and `isZeroAddress` functions. These functions are wrappers
for the functions provided by `etheres.js`.
