+++ src/web3/hooks/useSubscribeToERC20TransferEvent.ts
@@ -3,7 +3,7 @@ import { BigNumber } from "@ethersproject/bignumber"
 import { Token } from "../../enums/token"
 import { useSubscribeToContractEvent } from "./useSubscribeToContractEvent"
 import { useTokenBalance } from "../../hooks/useTokenBalance"
-import { isSameETHAddress } from "../../utils/isSameETHAddress"
+import { isSameETHAddress } from "../../web3/utils"
 import { useTokenState } from "../../hooks/useTokenState"
 import { useToken } from "../../hooks/useToken"
 import { useCallback } from "react"
