+++ src/web3/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./multicall"
 export * from "./events"
+export * from "./address"
